Creating file for utils
