+++ lyncs_io/formats.py
@@ -5,11 +5,10 @@ __all__ = [
     "formats",
 ]
 
-from functools import wraps
-from io import IOBase
 import pickle
 import json
 from .format import Format, Formats
+from .utils import swap, open_file
 
 formats = Formats()
 
@@ -20,28 +19,6 @@ def register(*names, **kwargs):
     fmt = Format(names[0], alias=names[1:], **kwargs)
     for name in names:
         formats[name.lower()] = fmt
-
-
-def swap(fnc):
-    "Returns a wrapper that swaps the first two arguments of the function"
-    return wraps(fnc)(
-        lambda fname, data, *args, **kwargs: fnc(data, fname, *args, **kwargs)
-    )
-
-
-def open_file(fnc, arg=0, flag="rb"):
-    "Returns a wrapper that opens the file (at position arg) if needed"
-
-    @wraps(fnc)
-    def wrapped(*args, **kwargs):
-        if len(args) <= arg:
-            raise ValueError(f"filename not found at position {arg}")
-        if isinstance(args[arg], IOBase):
-            return fnc(*args, **kwargs)
-        args = list(args)
-        with open(args[arg], flag) as fptr:
-            args[arg] = fptr
-            return fnc(*args, **kwargs)
 
 
 register(
