+++ lyncs_io/utils.py
@@ -1,0 +1,28 @@
+"""
+Function utils
+"""
+
+from functools import wraps
+from io import IOBase
+
+
+def swap(fnc):
+    "Returns a wrapper that swaps the first two arguments of the function"
+    return wraps(fnc)(
+        lambda fname, data, *args, **kwargs: fnc(data, fname, *args, **kwargs)
+    )
+
+
+def open_file(fnc, arg=0, flag="rb"):
+    "Returns a wrapper that opens the file (at position arg) if needed"
+
+    @wraps(fnc)
+    def wrapped(*args, **kwargs):
+        if len(args) <= arg:
+            raise ValueError(f"filename not found at position {arg}")
+        if isinstance(args[arg], IOBase):
+            return fnc(*args, **kwargs)
+        args = list(args)
+        with open(args[arg], flag) as fptr:
+            args[arg] = fptr
+            return fnc(*args, **kwargs)
